contacts responsiveness
