+++ src/pages/styling/Home.css
@@ -48,11 +48,23 @@ h2 {
 }
 
 .about-me {
-  margin: 5% 0 !important;
+  margin: 7% 0 !important;
   text-align: left
 }
 
 /* responsiveness */
+
+@media (min-width: 1600px) and (max-width: 2200px) {
+  /* Styles for large devices */
+  .home-section {
+    min-height: 100vh;
+    /* width: 1500px; */
+    /* margin: auto; */
+  }
+
+}
+
+
 @media (max-width: 768px) {
   .header-container {
     flex-direction: column; /* On smaller screens, switch to a column layout */
