+++ src/pages/styling/ProjectGallery.css
@@ -8,3 +8,14 @@ h1 {
   margin: 2.5rem 0;
   text-align: left;
 }
+
+
+/* responsiveness */
+@media (min-width: 1600px) and (max-width: 2200px) {
+  /* Styles for large devices */
+  h1 {
+    margin-top: 4.5rem;
+
+  }
+
+}
